requirements.txt
streamlit==1.41.1
pandas==2.2.3
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
plotly==5.24.1
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
scikit-learn==1.6.0
sentence-transformers==3.3.1
numpy==2.2.1
python-dotenv==1.0.1
apscheduler==3.10.4
playwright==1.49.1
requests==2.32.3
